+++ src/lib/account/index.ts
@@ -1,6 +1,7 @@
 export * from "./types";
 export * from "./client";
 export * from "./store";
+export * from "./descriptor";
 export * from "./dispatch";
 export * from "./hosted-vault";
 export * from "./provenance";